Move cannon subsystem and relevant commands to cannon package.
Move the cannon subsystem and the relevant commands to that subsystem to have
nice organization. There might be changes yet, as it is being debated yet if
there should be sub-subsystem for the different things that make up the very
diverse cannon.
