+++ src/org/robockets/tshirtcannon/RobotMap.java
@@ -1,7 +1,8 @@
 package org.robockets.tshirtcannon;
 
 import edu.wpi.first.wpilibj.Solenoid;
-import org.robockets.tshirtcannon.subsystems.Cannon;
+
+import org.robockets.tshirtcannon.subsystems.cannon.Cannon;
 
 /**
  * The RobotMap is a mapping from the ports sensors and actuators are wired into
